koord-scheduler: optimize NodeNUMAResource scoring
